Move uninstall integration test into its own package
